🚚(front) move DashboardVideoLiveInfoModal
Move DashboardVideoLiveInfoModal in the DashboardVideoLiveControlPane/customs
folder. Indeed, this component is only used by the
DashboardVideoLiveControlPane.
